Moves SWA auth model into shared project
