+++ src/Web/Authentication/StaticWebAppsAuthStateProvider.cs
@@ -1,7 +1,7 @@
 using System.Security.Claims;
 using System.Text.Json;
 using Microsoft.AspNetCore.Components.Authorization;
-using RunSuggestion.Web.Authentication.Models;
+using RunSuggestion.Shared.Models.Auth;
 
 namespace RunSuggestion.Web.Authentication;
